feat: 인벤토리 추가 보상 테스트 완료
