+++ Assets/Scripts/Offering/OfferingData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cc31ed61bb16775429ad87ce66339ee1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
