+++ Assets/Scripts/Offering/OfferingSetting.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: cc31ed61bb16775429ad87ce66339ee1, type: 3}
+  m_Name: OfferingSetting
+  m_EditorClassIdentifier: 
+  tier1Presents:
+  - cthulhuPresent: Expand Inventory +1
+  tier2Presents: []
+  tier3Presents: []
