+++ Assets/Scripts/Offering/OfferingSetting.asset.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 138b7cd346b044c4c9b02909f1dc2024
-DefaultImporter:
+guid: 4369b28d8239b974e81d176515371f80
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
